pytest: Separate coltypes tests
